application/production_requirements.txt
annotated-types==0.6.0
aiosqlite==0.20.0
anyio==4.3.0
asyncpg==0.29.0
attrs==23.2.0
bleach==6.1.0
click==8.1.7
fastapi==0.110.1
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
iniconfig==2.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==24.0
pathspec==0.12.1
pbr==6.0.0
pillow==10.3.0
platformdirs==4.2.0
pluggy==1.4.0
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
setuptools==69.2.0
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
SQLAlchemy==2.0.29
starlette==0.37.2
stevedore==5.2.0
typing_extensions==4.11.0
uvicorn==0.29.0
uvloop==0.19.0
webencodings==0.5.1
websockets==12.0
wheel==0.43.0
MarkupSafe==2.1.5
Mako==1.3.3
alembic==1.13.1